requirements.txt
Flask>=3.1,<3.2
Flask-SQLAlchemy>=3.1,<3.2
Flask-Migrate>=4.1,<4.2
Flask-Login>=0.6.3,<0.7
Flask-WTF>=1.2,<1.3
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<2.0
email-validator>=2.2,<3.0
pytest>=8.3,<9.0